refactor(task-lifecycle): finish helper extraction follow-ups
