+++ src/orchestrator/execution/__tests__/task-lifecycle-cycle.test.ts
@@ -391,427 +391,4 @@ describe("TaskLifecycle", async () => {
     });
   });
 
-  // ─────────────────────────────────────────────
-  // Persistence
-  // ─────────────────────────────────────────────
-
-  describe("persistence", async () => {
-    it("verification result saved to correct path", async () => {
-      const llm = createMockLLMClient([LLM_REVIEW_PASS]);
-      const lifecycle = createLifecycle(llm);
-      const task = makeTask({ id: "task-persist-test" });
-      const result: import("../task/task-lifecycle.js").AgentResult = {
-        success: true,
-        output: "done",
-        error: null,
-        exit_code: 0,
-        elapsed_ms: 100,
-        stopped_reason: "completed",
-      };
-
-      await lifecycle.verifyTask(task, result);
-
-      const saved = await stateManager.readRaw("verification/task-persist-test/verification-result.json");
-      expect(saved).not.toBeNull();
-      expect((saved as Record<string, unknown>).task_id).toBe("task-persist-test");
-    });
-
-    it("task history accumulates entries", async () => {
-      const llm = createMockLLMClient([]);
-      const lifecycle = createLifecycle(llm);
-
-      // Simulate two pass verdicts
-      for (let i = 1; i <= 2; i++) {
-        const task = makeTask({ id: `task-${i}` });
-        const vr: import("../../../base/types/task.js").VerificationResult = {
-          task_id: `task-${i}`,
-          verdict: "pass",
-          confidence: 0.9,
-          evidence: [
-            { layer: "independent_review", description: "OK", confidence: 0.8 },
-          ],
-          dimension_updates: [],
-          timestamp: new Date().toISOString(),
-        };
-
-        await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-        await lifecycle.handleVerdict(task, vr);
-      }
-
-      const history = await stateManager.readRaw("tasks/goal-1/task-history.json") as Array<Record<string, unknown>>;
-      expect(history.length).toBe(2);
-      expect(history[0]!.task_id).toBe("task-1");
-      expect(history[1]!.task_id).toBe("task-2");
-    });
-
-    it("task history records primary_dimension", async () => {
-      const llm = createMockLLMClient([]);
-      const lifecycle = createLifecycle(llm);
-      const task = makeTask({ primary_dimension: "coverage" });
-      const vr: import("../../../base/types/task.js").VerificationResult = {
-        task_id: "task-1",
-        verdict: "pass",
-        confidence: 0.9,
-        evidence: [
-          { layer: "independent_review", description: "OK", confidence: 0.8 },
-        ],
-        dimension_updates: [],
-        timestamp: new Date().toISOString(),
-      };
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      await lifecycle.handleVerdict(task, vr);
-
-      const history = await stateManager.readRaw("tasks/goal-1/task-history.json") as Array<Record<string, unknown>>;
-      expect(history[0]!.primary_dimension).toBe("coverage");
-    });
-
-    it("task history records consecutive_failure_count on failure", async () => {
-      const llm = createMockLLMClient([]);
-      const lifecycle = createLifecycle(llm);
-      const task = makeTask({ consecutive_failure_count: 1 });
-      const vr: import("../../../base/types/task.js").VerificationResult = {
-        task_id: "task-1",
-        verdict: "fail",
-        confidence: 0.9,
-        evidence: [
-          { layer: "independent_review", description: "Direction OK", confidence: 0.8 },
-          { layer: "self_report", description: "Tried", confidence: 0.3 },
-        ],
-        dimension_updates: [],
-        timestamp: new Date().toISOString(),
-      };
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      await lifecycle.handleFailure(task, vr);
-
-      const history = await stateManager.readRaw("tasks/goal-1/task-history.json") as Array<Record<string, unknown>>;
-      expect(history[0]!.consecutive_failure_count).toBe(2);
-    });
-
-    it("executeTask persists running state before execution", async () => {
-      const llm = createMockLLMClient([]);
-      const lifecycle = createLifecycle(llm);
-      let statusDuringExecution = "";
-      const adapter: import("../task/task-lifecycle.js").IAdapter = {
-        adapterType: "mock",
-        async execute() {
-          // Check status during execution
-          const raw = await stateManager.readRaw("tasks/goal-1/task-1.json") as Record<string, unknown>;
-          statusDuringExecution = raw?.status as string;
-          return {
-            success: true, output: "ok", error: null,
-            exit_code: 0, elapsed_ms: 10, stopped_reason: "completed" as const,
-          };
-        },
-      };
-      const task = makeTask();
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      await lifecycle.executeTask(task, adapter);
-
-      expect(statusDuringExecution).toBe("running");
-    });
-  });
-
-  // ─────────────────────────────────────────────
-  // failure handling paths
-  // ─────────────────────────────────────────────
-
-  describe("failure handling paths", async () => {
-    // ── Test 1: L1 mechanical verification ─────────────────────────────────
-
-    it("L1 mechanical criteria detected: evidence includes mechanical layer", async () => {
-      // Task has a shell-command verification method → L1 applicable (assumed pass).
-      // L2 returns "pass" → overall verdict is "pass".
-      const llm = createMockLLMClient([
-        LLM_REVIEW_PASS, // L2 review
-      ]);
-      const lifecycle = createLifecycle(llm);
-      const task = makeTask({
-        id: "task-l1-mech",
-        success_criteria: [
-          {
-            description: "Tests pass",
-            verification_method: "npm test",
-            is_blocking: true,
-          },
-        ],
-      });
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      const result = await lifecycle.verifyTask(task, {
-        success: true,
-        output: "All tests passed",
-        error: null,
-        exit_code: 0,
-        elapsed_ms: 50,
-        stopped_reason: "completed",
-      });
-
-      // Evidence should include the mechanical layer
-      const layers = result.evidence.map((e) => e.layer);
-      expect(layers).toContain("mechanical");
-      expect(layers).toContain("independent_review");
-      // L1 assumed pass + L2 pass → "pass"
-      expect(result.verdict).toBe("pass");
-    });
-
-    it("L1 mechanical criteria detected + L2 fail → re-review → overall fail", async () => {
-      // L1 assumed pass + L2 fail → triggers re-review; if re-review also fails → "fail"
-      const llm = createMockLLMClient([
-        LLM_REVIEW_FAIL, // first L2 review
-        LLM_REVIEW_FAIL, // re-review
-      ]);
-      const lifecycle = createLifecycle(llm);
-      const task = makeTask({
-        id: "task-l1-mech-fail",
-        success_criteria: [
-          {
-            description: "Build succeeds",
-            verification_method: "npx tsc --noEmit",
-            is_blocking: true,
-          },
-        ],
-      });
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      const result = await lifecycle.verifyTask(task, {
-        success: false,
-        output: "TypeScript errors found",
-        error: "Compilation failed",
-        exit_code: 1,
-        elapsed_ms: 30,
-        stopped_reason: "error",
-      });
-
-      expect(result.verdict).toBe("fail");
-    });
-
-    it("L1 not applicable (no shell command): evidence has no mechanical layer, confidence 0.6 on pass", async () => {
-      const llm = createMockLLMClient([
-        LLM_REVIEW_PASS, // L2 returns pass
-      ]);
-      const lifecycle = createLifecycle(llm);
-      const task = makeTask({
-        id: "task-l1-skip",
-        success_criteria: [
-          {
-            description: "Peer review approved",
-            verification_method: "Manual code review",
-            is_blocking: true,
-          },
-        ],
-      });
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      const result = await lifecycle.verifyTask(task, {
-        success: true,
-        output: "Review done",
-        error: null,
-        exit_code: 0,
-        elapsed_ms: 20,
-        stopped_reason: "completed",
-      });
-
-      // L1 skipped → pass with lower confidence (0.6)
-      expect(result.verdict).toBe("pass");
-      expect(result.confidence).toBe(0.6);
-      const layers = result.evidence.map((e) => e.layer);
-      expect(layers).not.toContain("mechanical");
-    });
-
-    // ── Test 2: keep / discard / escalate paths ─────────────────────────────
-
-    it("partial verdict with direction correct (partial = direction correct) → action is keep", async () => {
-      const llm = createMockLLMClient([]);
-      const lifecycle = createLifecycle(llm);
-      const task = makeTask({ id: "task-keep" });
-
-      const vr: import("../../../base/types/task.js").VerificationResult = {
-        task_id: "task-keep",
-        verdict: "partial",
-        confidence: 0.5,
-        evidence: [
-          { layer: "independent_review", description: "Partial progress", confidence: 0.6 },
-        ],
-        dimension_updates: [
-          { dimension_name: "dim", previous_value: 0.5, new_value: 0.65, confidence: 0.5 },
-        ],
-        timestamp: new Date().toISOString(),
-      };
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      // Save a goal so dimension updates can be applied
-      await stateManager.writeRaw("goals/goal-1.json", {
-        id: "goal-1",
-        dimensions: [{ name: "dim", current_value: 0.5 }],
-      });
-
-      const result = await lifecycle.handleVerdict(task, vr);
-      expect(result.action).toBe("keep");
-    });
-
-    it("fail verdict with reversible task → revert succeeds → action is discard", async () => {
-      // handleFailure: fail verdict → direction wrong (verdict="fail"), reversible →
-      // attemptRevert → revert succeeds → discard.
-      const llm = createMockLLMClient([
-        REVERT_SUCCESS, // attemptRevert calls llm.sendMessage
-      ]);
-      const lifecycle = createLifecycle(llm);
-      const task = makeTask({
-        id: "task-discard",
-        reversibility: "reversible",
-        consecutive_failure_count: 0,
-      });
-
-      const vr: import("../../../base/types/task.js").VerificationResult = {
-        task_id: "task-discard",
-        verdict: "fail",
-        confidence: 0.9,
-        evidence: [
-          { layer: "independent_review", description: "Nothing worked", confidence: 0.8 },
-        ],
-        dimension_updates: [],
-        timestamp: new Date().toISOString(),
-      };
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      const result = await lifecycle.handleFailure(task, vr);
-      expect(result.action).toBe("discard");
-    });
-
-    it("consecutive_failure_count reaches 3 → action is escalate", async () => {
-      const llm = createMockLLMClient([]);
-      const lifecycle = createLifecycle(llm);
-      // Already at 2; handleFailure will increment to 3 → escalate before direction check
-      const task = makeTask({
-        id: "task-escalate",
-        consecutive_failure_count: 2,
-        reversibility: "reversible",
-      });
-
-      const vr: import("../../../base/types/task.js").VerificationResult = {
-        task_id: "task-escalate",
-        verdict: "fail",
-        confidence: 0.9,
-        evidence: [
-          { layer: "independent_review", description: "Repeated failures", confidence: 0.8 },
-        ],
-        dimension_updates: [],
-        timestamp: new Date().toISOString(),
-      };
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      const result = await lifecycle.handleFailure(task, vr);
-      expect(result.action).toBe("escalate");
-      expect(result.task.consecutive_failure_count).toBe(3);
-    });
-
-    it("fail verdict with irreversible task and direction wrong → action is escalate without revert", async () => {
-      // No LLM response needed — irreversible goes straight to escalate
-      const llm = createMockLLMClient([]);
-      const lifecycle = createLifecycle(llm);
-      const task = makeTask({
-        id: "task-irreversible-fail",
-        reversibility: "irreversible",
-        consecutive_failure_count: 0,
-      });
-
-      const vr: import("../../../base/types/task.js").VerificationResult = {
-        task_id: "task-irreversible-fail",
-        verdict: "fail",
-        confidence: 0.9,
-        evidence: [
-          { layer: "independent_review", description: "Did not meet criteria", confidence: 0.8 },
-        ],
-        dimension_updates: [],
-        timestamp: new Date().toISOString(),
-      };
-
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-      const result = await lifecycle.handleFailure(task, vr);
-      expect(result.action).toBe("escalate");
-    });
-
-    // ── Test 3: Adapter timeout ─────────────────────────────────────────────
-
-    it("adapter that rejects with an Error → executeTask catches and returns error result", async () => {
-      const llm = createMockLLMClient([]);
-      const lifecycle = createLifecycle(llm);
-
-      const timeoutAdapter: import("../task/task-lifecycle.js").IAdapter = {
-        adapterType: "mock-timeout",
-        async execute() {
-          throw new Error("Adapter execution timed out after 30000ms");
-        },
-      };
-
-      const task = makeTask({ id: "task-timeout" });
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-
-      const result = await lifecycle.executeTask(task, timeoutAdapter);
-
-      // executeTask catches the error and returns a graceful failure result
-      expect(result.success).toBe(false);
-      expect(result.stopped_reason).toBe("error");
-      expect(result.error).toContain("timed out");
-    });
-
-    it("adapter timeout followed by verifyTask produces a fail verdict", async () => {
-      // When adapter throws, execution result is {success: false}.
-      // verifyTask then calls LLM review with this failed output.
-      const llm = createMockLLMClient([
-        LLM_REVIEW_FAIL, // L2 review for the failed execution
-      ]);
-      const lifecycle = createLifecycle(llm);
-
-      const timeoutAdapter: import("../task/task-lifecycle.js").IAdapter = {
-        adapterType: "mock-timeout",
-        async execute() {
-          throw new Error("Connection timeout");
-        },
-      };
-
-      const task = makeTask({
-        id: "task-timeout-verify",
-        success_criteria: [
-          {
-            description: "Deployment successful",
-            verification_method: "Check deployment status manually",
-            is_blocking: true,
-          },
-        ],
-      });
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-
-      const executionResult = await lifecycle.executeTask(task, timeoutAdapter);
-      expect(executionResult.success).toBe(false);
-
-      const verificationResult = await lifecycle.verifyTask(task, executionResult);
-      expect(verificationResult.verdict).toBe("fail");
-    });
-
-    it("adapter that throws a non-Error value is handled gracefully", async () => {
-      const llm = createMockLLMClient([]);
-      const lifecycle = createLifecycle(llm);
-
-      const badAdapter: import("../task/task-lifecycle.js").IAdapter = {
-        adapterType: "mock-bad",
-        async execute() {
-          // eslint-disable-next-line @typescript-eslint/no-throw-literal
-          throw "string error value";
-        },
-      };
-
-      const task = makeTask({ id: "task-bad-throw" });
-      await stateManager.writeRaw(`tasks/${task.goal_id}/${task.id}.json`, task);
-
-      const result = await lifecycle.executeTask(task, badAdapter);
-
-      expect(result.success).toBe(false);
-      expect(result.stopped_reason).toBe("error");
-      expect(typeof result.error).toBe("string");
-    });
-  });
 });
